# Fix string.split and array.join performance