--- a/Jint/Native/Array/ArrayExecutionContext.cs
+++ b/Jint/Native/Array/ArrayExecutionContext.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Text;
 using System.Threading;
 
 namespace Jint.Native.Array
@@ -15,6 +16,7 @@ namespace Jint.Native.Array
         private JsValue[] _callArray1;
         private JsValue[] _callArray3;
         private JsValue[] _callArray4;
+        private StringBuilder _stringBuilder;
 
         private ArrayExecutionContext()
         {
@@ -25,6 +27,7 @@ namespace Jint.Native.Array
         public JsValue[] CallArray1 => _callArray1 = _callArray1 ?? new JsValue[1];
         public JsValue[] CallArray3 => _callArray3 = _callArray3 ?? new JsValue[3];
         public JsValue[] CallArray4 => _callArray4 = _callArray4 ?? new JsValue[4];
+        public StringBuilder StringBuilder => _stringBuilder = _stringBuilder ?? new StringBuilder();
 
         public static ArrayExecutionContext Current => _executionContext.Value;
     }