… Jint.Benchmark/Scripts/dromaeo/dromaeo-string-base64.js
… Jint/Native/Array/ArrayPrototype.cs